Fire interceptor prior to new status
If the interceptor tries to salvage the page by loading something else
(like in the case of the BlogCFC compat SES URL module) it has to
manually set the status code BACK to 200.  Waiting to change the status
code to 404 until after the interceptor has had a whack at it removes
that burden from the interceptor.