set JRE_HOME only if it exists
currently the JRE_HOME is set regardless of whether it exists (Express with JRE) or not (Express without JRE), making the no-JRE version unusable as it overwrites the JRE_HOME variable if set properly in the Environment.  I suspect that the *nix version .sh suffers from the same issue, but have not checked that as of yet.