CONTENTBOX-1085
Fixes #1085 
Ensures that the setting name ( slug ) is unique, and cannot be duplicated. 
Without this, if a weird cache race condition is found, the Preflight check would insert 135 duplicate settings, causing every ORM findWhere to return more than 1 record for a setting name, stopping ColdBox from Initiliazing, resulting in many strange and not so wonderful errors. 

https://ortussolutions.atlassian.net/browse/CONTENTBOX-1085